+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## **WORK IN PROGRESS**
 
 * (theimo1221) Fix logging in weather service
+* (theimo1221) Reduce Daikin Logging in success case
 
 ## 1.0.18-alpha.0 (2022-07-24)